Run tests in pipeline
